Address #964, fix critical bug in RLE2 decoding.
This is really stupid and obvious in hindsight, spans > 127 were
interepted incorrectly due to integer promotion.

+++ src/common/net-pack.c
@@ -719,7 +719,7 @@ int cv_decode_rle2(cave_view_type* dst, cq* src, int len) {
 		if (a == 0xFF)
 		{
 			/* Get the number of repetitions */
-			n = c;
+			n = (byte)c;
 
 			/* Is it even legal? */
 			if (x + n > len)
